refactor: change proto package to models
