+++ plugins/extractors/metabase/metabase_test.go
@@ -15,9 +15,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/odpf/meteor/models/odpf/assets"
 	"github.com/odpf/meteor/plugins"
 	"github.com/odpf/meteor/plugins/extractors/metabase"
-	"github.com/odpf/meteor/proto/odpf/assets"
 	"github.com/odpf/meteor/test"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
