+++ plugins/extractors/mssql/mssql.go
@@ -11,9 +11,9 @@ import (
 	"github.com/odpf/meteor/plugins"
 	"github.com/odpf/meteor/registry"
 
-	"github.com/odpf/meteor/proto/odpf/assets"
-	"github.com/odpf/meteor/proto/odpf/assets/common"
-	"github.com/odpf/meteor/proto/odpf/assets/facets"
+	"github.com/odpf/meteor/models/odpf/assets"
+	"github.com/odpf/meteor/models/odpf/assets/common"
+	"github.com/odpf/meteor/models/odpf/assets/facets"
 	"github.com/odpf/meteor/utils"
 )
 
@@ -43,7 +43,7 @@ var sampleConfig = `
 type Extractor struct {
 	out         chan<- interface{}
 	excludedDbs map[string]bool
-	logger log.Logger
+	logger      log.Logger
 }
 
 // New returns a pointer to an initialized Extractor Object
