+++ plugins/extractors/postgres/postgres_test.go
@@ -12,9 +12,9 @@ import (
 	"database/sql"
 
 	_ "github.com/lib/pq"
+	"github.com/odpf/meteor/models/odpf/assets"
 	"github.com/odpf/meteor/plugins"
 	"github.com/odpf/meteor/plugins/extractors/postgres"
-	"github.com/odpf/meteor/proto/odpf/assets"
 	"github.com/odpf/meteor/test"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
